screens: preserve the landfill pointer at legend blits
Forward-port 0790818a. place_2x2_block takes an unsigned-char pointer; remove the stale 32-bit int cast that survived the earlier landfill-base correction. The rebuilt PS.EXE remains whole-file byte-exact with zero fixup target mismatches.

+++ src/screens.c
@@ -2845,7 +2845,7 @@ void place_legend_block(int sprite_idx, int x, int y)
                  + (landfill[sprite_idx * 16 + 0xD] << 8);
     for (yo = 0; yo < 16; yo += 2) {
         for (xo = 0; xo < 16; xo += 2) {
-            place_2x2_block((int)landfill + sprite_start,
+            place_2x2_block(landfill + sprite_start,
                             (x + xo) + (y + yo) * screen_width);
         }
     }
